+++ src/main/java/org/openlmis/fulfillment/service/referencedata/UserDto.java
@@ -38,6 +38,7 @@ public class UserDto extends BaseDto {
   private boolean verified;
   private boolean active;
   private boolean loginRestricted;
+  private boolean allowNotify;
   private UUID homeFacilityId;
   private Set<RoleAssignmentDto> roleAssignments;